+++ html/muscle-by-example/examples/stringmatcher/example_2_regex_matching.cpp
@@ -50,7 +50,7 @@ int main(int argc, char ** argv)
             matchCount++;
          }
       }
-      printf("%i/%u dictionary words matched regex pattern [%s]\n", matchCount, ARRAYITEMS(_commonWords), s());
+      printf("%i/%u dictionary words matched regex pattern [%s]\n", matchCount, (unsigned int) ARRAYITEMS(_commonWords), s());
       printf("\n");
    }
